+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ CHANGELOG
 * Add backward compatibility for alfacase files which does not specify the version in plugins.
 * Add ``version`` property for ``PluginsDescription``
 * **Breaking Change**: Add ``alfasim_get_version`` GUI specification hook to be implemented by plugins.
+* **Breaking Change**: ``HydrodynamicModelType.ThreeLayersNoBubbleGasOilWater``  renamed to ``HydrodynamicModelType.TwoLayersGasOilWater``
 
 1.1.0 (2025-05-28)
 ==================